Centralize `npm` stuff in root
Since currently `npm` commands apply globally and not to single gems.
