+++ .rubocop.yml
@@ -1214,9 +1214,8 @@ RSpec/ContextWording:
 RSpec/DescribeClass:
   Exclude:
     - spec/bundle_spec.rb
+    - spec/bundles_spec.rb
     - spec/i18n_spec.rb
-    - decidim-admin/spec/bundles_spec.rb
-    - decidim-comments/spec/bundle_spec.rb
     - decidim-core/spec/lib/global_engines_spec.rb
     - "**/tasks/**/*"
